requirements.txt
scikit-learn
tensorflow
pillow
plotly
numpy
kaleido
